replace kmalloc size uint32_t with size_t

+++ include/kernel/mm.h
@@ -4,7 +4,7 @@
 #include <lib/stddef.h>
 
 void init_mem(uint32_t start, uint32_t end);
-void *kmalloc(uint32_t size);
+void *kmalloc(size_t size);
 void kfree(void *p);
 
 #endif
